requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Mail==0.9.1
email-validator==2.1.0.post1
Werkzeug==3.0.1
SQLAlchemy==2.0.23
WTForms==3.1.1
PyJWT==2.8.0
redis==5.0.1
requests==2.31.0
oauthlib==3.2.2
python-dotenv==1.0.0
gunicorn==21.2.0
bcrypt==4.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.4
urllib3==2.1.0
